+++ packages/react-native/src/index.ts
@@ -31,3 +31,4 @@ export * from './types/Notification';
 export * from './types/Trigger';
 export * from './types/NotificationIOS';
 export * from './types/NotificationAndroid';
+export type { FcmRemoteMessage, FcmConfig } from './fcm/types';